Add routes and server.js
