+++ controllers/homeController.js
@@ -1,5 +1,9 @@
+const source = require('rfr');
+
+const Bot = source('lib/bot');
+
 const HomeController = {
-    get(req, res, Bot) {
+    get(req, res) {
         res.render('home', {
             clientID: Bot.client.user.id,
         });
